+++ app/containers/Transactions/SendOnePage.js
@@ -6,6 +6,7 @@ import { bindActionCreators } from 'redux';
 import {
   cleanForm,
   inputChanged,
+  validateInput,
   validateForm,
   transactionAccepted,
   transactionRejected,
@@ -99,6 +100,7 @@ export default withRouter(connect(
       {
         cleanForm,
         inputChanged,
+        validateInput,
         validateForm,
         transactionAccepted,
         transactionRejected,