+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+# [0.4.2] - 2021-06-08
+### Added
+- Gateway fee threshold
+### Fixed
+- Gateway form validation
+
 # [0.4.1] - 2021-05-06
 ### Fixed
 - Undefined gateways error
@@ -74,8 +80,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Storage encryption
 - Docs
 
-[Unreleased]: https://github.com/adshares/ads-browser-wallet/compare/v0.4.1...HEAD
+[Unreleased]: https://github.com/adshares/ads-browser-wallet/compare/v0.4.2...HEAD
 
+[0.4.2]: https://github.com/adshares/ads-browser-wallet/compare/v0.4.1...v0.4.2
 [0.4.1]: https://github.com/adshares/ads-browser-wallet/compare/v0.4.0...v0.4.1
 [0.4.0]: https://github.com/adshares/ads-browser-wallet/compare/v0.3.2...v0.4.0
 [0.3.2]: https://github.com/adshares/ads-browser-wallet/compare/v0.3.1...v0.3.2